fix: keep docker setup helpers outside component
