+++ src/components/chat/input/SandboxModeSwitcher.tsx
@@ -5,7 +5,8 @@ import { useClickOutside } from "@/hooks/useClickOutside"
 import { cn } from "@/lib/utils"
 import { getTransport } from "@/lib/transport-provider"
 import { logger } from "@/lib/logger"
-import { DockerSetupHint, type DockerStatus } from "@/components/settings/DockerSetupHint"
+import { DockerSetupHint } from "@/components/settings/DockerSetupHint"
+import type { DockerStatus } from "@/components/settings/dockerSetup"
 import { Box, Copy, Folder, Shield, ShieldCheck } from "lucide-react"
 import type { SandboxMode } from "@/types/chat"
